fixed prominent inconsistency of ids of connected entities
On adding a new entity, the UI generates a temporary UID for it.
The API then returns the added entity with a (possibly) different UID.
If the entity to be connected to other entities, the UI in some cases
used UI-generated UIDs instead of API-generated ones. So far, those,
UIDs were the same but this might change in the future leading to storing
non-existent (ui-generated) UIDs to the arrays of connected entities.

The fix changed the behaviour of UI so that it always uses API-generated
UIDs.